# Architecture_DZ_10/API/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cloud service for robot-vacuum cleaner",
    "description": "This is a study project considering server based on the OpenAPI 3.0 specification. \nYou can now help us improve the API whether it's by making changes to the definition itself or to the code.\nThat way, with time, we can improve the API in general, and expose some of the new features in OAS3.",
    "contact": {
      "email": "kardan97@yandex.ru"
    },
    "version": "0.0.2"
  },
  "servers": [
    {
      "url": "https://petstore3.swagger.io/api/v3"
    },
    {
      "url": "https://localhost:8080/api/v1"
    }
  ],
...
  "paths": {
...
    "/timetable/newclean": {
      "post": {
        "tags": [
          "timetable"
        ],
        "summary": "Create a new record in timetable",
        "description": "Create a new record in timetable",
        "operationId": "newRecord",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Timetable"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Timetable"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Timetable"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timetable"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/timetable/all": {
      "get": {
        "tags": [
          "timetable"
        ],
        "summary": "Returns timetable with all known cleanings",
        "description": "Returns timetable with all known cleanings",
        "operationId": "getTimetable",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timetable"
                }
              }
            }
          },
          "404": {
            "description": "Timetable not found"
          }
        }
      },
      "delete": {
        "tags": [
          "timetable"
        ],
        "summary": "Delete all records from timetable",
        "operationId": "clearTimetable",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timetable"
                }
              }
            }
          }
        }
      }
    },
    "/timetable/{recordId}": {
      "get": {
        "tags": [
          "timetable"
        ],
        "summary": "Get record in timetable by ID",
        "operationId": "getRecord",
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timetable"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Timetable"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Record not found in timetable"
          }
        }
      },
      "put": {
        "tags": [
          "timetable"
        ],
        "summary": "Update record in timetable by ID",
        "operationId": "editRecord",
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timetable"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Timetable"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Record not found in timetable"
          }
        }
      },
      "delete": {
        "tags": [
          "timetable"
        ],
        "summary": "Delete record from timetable by ID",
        "operationId": "deleteRecord",
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Record not found in timetable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Timetable": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "datetime": {
            "type": "string",
            "format": "date-time"
          },
          "roomID": {
            "$ref": "#/components/schemas/Room"
          },
          "resultID": {
            "$ref": "#/components/schemas/Result"
          }
        },
        "xml": {
          "name": "timetable"
        }
      },
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roomname": {
            "type": "string"
          },
          "roomsquare": {
            "type": "string"
          }
        },
        "xml": {
          "name": "room"
        }
      },
      "Result": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "Details": {
            "type": "string",
            "description": "What does this code mean"
          }
        },
        "xml": {
          "name": "result"
        }
      },
...
    },
...
  }
}